Fix bad custom animation example
The custom animation example suggests that a stack-allocated
AnimationImplementation variable can be used, but this is incorrect.
The variable is not copied and so must remain valid while the
animation runs.

Rework the example to use a global variable for the implementation.

Signed-off-by: Lincoln Ramsay <a1291762@gmail.com>
(cherry picked from commit 967f7166a09cfca03783b1778fa2dd0ec17640b6)